+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.DataLayer/DataLayerService.HorizontalAxis.cs
@@ -7,27 +7,27 @@ namespace Ferretto.VW.MAS.DataLayer
     {
         #region Properties
 
-        public decimal AntiClockWiseRun => this.GetDecimalConfigurationValue((long)HorizontalAxis.AntiClockWiseRun, ConfigurationCategory.HorizontalAxis);
+        public decimal AntiClockWiseRun => this.GetDecimalConfigurationValue(HorizontalAxis.AntiClockWiseRun, ConfigurationCategory.HorizontalAxis);
 
-        public decimal ClockWiseRun => this.GetDecimalConfigurationValue((long)HorizontalAxis.ClockWiseRun, ConfigurationCategory.HorizontalAxis);
+        public decimal ClockWiseRun => this.GetDecimalConfigurationValue(HorizontalAxis.ClockWiseRun, ConfigurationCategory.HorizontalAxis);
 
-        public bool HomingExecutedHA => this.GetBoolConfigurationValue((long)HorizontalAxis.HomingExecuted, ConfigurationCategory.HorizontalAxis);
+        public bool HomingExecutedHA => this.GetBoolConfigurationValue(HorizontalAxis.HomingExecuted, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxEmptyAccelerationHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxEmptyAcceleration, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxEmptyAccelerationHA => this.GetDecimalConfigurationValue(HorizontalAxis.MaxEmptyAcceleration, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxEmptyDecelerationHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxEmptyDeceleration, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxEmptyDecelerationHA => this.GetDecimalConfigurationValue(HorizontalAxis.MaxEmptyDeceleration, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxEmptySpeedHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxEmptySpeed, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxEmptySpeedHA => this.GetDecimalConfigurationValue(HorizontalAxis.MaxEmptySpeed, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxFullAccelerationHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxFullAcceleration, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxFullAccelerationHA => this.GetDecimalConfigurationValue(HorizontalAxis.MaxFullAcceleration, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxFullDecelerationHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxFullDeceleration, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxFullDecelerationHA => this.GetDecimalConfigurationValue(HorizontalAxis.MaxFullDeceleration, ConfigurationCategory.HorizontalAxis);
 
-        public decimal MaxFullSpeed => this.GetDecimalConfigurationValue((long)HorizontalAxis.MaxFullSpeed, ConfigurationCategory.HorizontalAxis);
+        public decimal MaxFullSpeed => this.GetDecimalConfigurationValue(HorizontalAxis.MaxFullSpeed, ConfigurationCategory.HorizontalAxis);
 
-        public decimal OffsetHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.Offset, ConfigurationCategory.HorizontalAxis);
+        public decimal OffsetHA => this.GetDecimalConfigurationValue(HorizontalAxis.Offset, ConfigurationCategory.HorizontalAxis);
 
-        public decimal ResolutionHA => this.GetDecimalConfigurationValue((long)HorizontalAxis.Resolution, ConfigurationCategory.HorizontalAxis);
+        public decimal ResolutionHA => this.GetDecimalConfigurationValue(HorizontalAxis.Resolution, ConfigurationCategory.HorizontalAxis);
 
         #endregion
     }
