+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.DataLayer/DataLayerService.VerticalManualMovements.cs
@@ -8,13 +8,13 @@ namespace Ferretto.VW.MAS.DataLayer
     {
         #region Properties
 
-        public decimal FeedRateAfterZero => this.GetDecimalConfigurationValue((long)VerticalManualMovements.FeedRateAfterZero, ConfigurationCategory.VerticalManualMovements);
+        public decimal FeedRateAfterZero => this.GetDecimalConfigurationValue(VerticalManualMovements.FeedRateAfterZero, ConfigurationCategory.VerticalManualMovements);
 
-        public decimal FeedRateVM => this.GetDecimalConfigurationValue((long)VerticalManualMovements.FeedRate, ConfigurationCategory.VerticalManualMovements);
+        public decimal FeedRateVM => this.GetDecimalConfigurationValue(VerticalManualMovements.FeedRate, ConfigurationCategory.VerticalManualMovements);
 
-        public decimal NegativeTargetDirection => this.GetDecimalConfigurationValue((long)VerticalManualMovements.NegativeTargetDirection, ConfigurationCategory.VerticalManualMovements);
+        public decimal NegativeTargetDirection => this.GetDecimalConfigurationValue(VerticalManualMovements.NegativeTargetDirection, ConfigurationCategory.VerticalManualMovements);
 
-        public decimal PositiveTargetDirection => this.GetDecimalConfigurationValue((long)VerticalManualMovements.PositiveTargetDirection, ConfigurationCategory.VerticalManualMovements);
+        public decimal PositiveTargetDirection => this.GetDecimalConfigurationValue(VerticalManualMovements.PositiveTargetDirection, ConfigurationCategory.VerticalManualMovements);
 
         #endregion
     }
